requirements.txt
Django>=4.2,<5.0
gunicorn>=21.2
python-decouple>=3.8
whitenoise>=6.6.0
psycopg2-binary>=2.9.9
django-tailwind>=3.5  # optional if you want Tailwind via Django app
dj-database-url>=0.5.0
django-cors-headers>=3.13.0
djangorest-framework>=3.14.0
djangocors-headers>=3.13.0